+++ components/partials/project/actionsCard/buttons/StarButton.tsx
